Constrain the rail chip row and split the chip converters (#898)
The chips row is now a Grid whose meta column takes the remaining width and trims, so the row fits the 310px rail even when every chip is long. The two vendor-chip converters move to their own files (one type per file), and the restating/misleading comments are dropped.
