+++ src/Capacitor.App/Views/SessionRailView.axaml
@@ -191,34 +191,37 @@
                                                                                                IsVisible="{Binding HasTitle}"
                                                                                                Foreground="{StaticResource KcapTextBrush}"
                                                                                                TextTrimming="CharacterEllipsis" />
-                                                                                    <!-- Every externally-sourced label is width-capped and trims, so a long
-                                                                                         vendor token, model id, or daemon name cannot overflow the 310px rail. -->
-                                                                                    <StackPanel Orientation="Horizontal" Spacing="6">
-                                                                                        <Border IsVisible="{Binding HasVendor}" CornerRadius="999" MaxWidth="120"
+                                                                                    <!-- A Grid, not a StackPanel: Meta takes the remaining (star) width so its
+                                                                                         ellipsis engages, and each chip is width-capped, so the whole row stays
+                                                                                         within the 310px rail even with long vendor/model/daemon names. Per-child
+                                                                                         left margins (not ColumnSpacing) so a hidden chip leaves no gap. -->
+                                                                                    <Grid ColumnDefinitions="Auto,Auto,Auto,*">
+                                                                                        <Border Grid.Column="0" IsVisible="{Binding HasVendor}" CornerRadius="999" MaxWidth="120"
                                                                                                 Padding="6,1" VerticalAlignment="Center"
                                                                                                 Background="{Binding Vendor, Converter={x:Static views:VendorChipBackgroundConverter.Instance}}">
                                                                                             <TextBlock Text="{Binding Vendor}" FontSize="10" FontWeight="SemiBold"
                                                                                                        TextTrimming="CharacterEllipsis"
                                                                                                        Foreground="{Binding Vendor, Converter={x:Static views:VendorChipForegroundConverter.Instance}}" />
                                                                                         </Border>
-                                                                                        <Border IsVisible="{Binding HasModel}" CornerRadius="999" MaxWidth="150"
-                                                                                                Padding="6,1" VerticalAlignment="Center"
+                                                                                        <Border Grid.Column="1" IsVisible="{Binding HasModel}" CornerRadius="999" MaxWidth="150"
+                                                                                                Padding="6,1" Margin="6,0,0,0" VerticalAlignment="Center"
                                                                                                 Background="{StaticResource KcapInfoDimBrush}">
                                                                                             <TextBlock Text="{Binding Model}" FontSize="10"
                                                                                                        Foreground="{StaticResource KcapInfoBrush}"
                                                                                                        TextTrimming="CharacterEllipsis" />
                                                                                         </Border>
-                                                                                        <Border IsVisible="{Binding MachineBadge, Converter={x:Static ObjectConverters.IsNotNull}}"
+                                                                                        <Border Grid.Column="2" IsVisible="{Binding MachineBadge, Converter={x:Static ObjectConverters.IsNotNull}}"
                                                                                                 Background="{StaticResource KcapSurfaceRaisedBrush}" MaxWidth="100"
-                                                                                                CornerRadius="3" Padding="4,0" VerticalAlignment="Center">
+                                                                                                CornerRadius="3" Padding="4,0" Margin="6,0,0,0" VerticalAlignment="Center">
                                                                                             <TextBlock Text="{Binding MachineBadge}" FontSize="10"
                                                                                                        TextTrimming="CharacterEllipsis"
                                                                                                        Foreground="{StaticResource KcapFaintBrush}" />
                                                                                         </Border>
-                                                                                        <TextBlock Text="{Binding Meta}" FontSize="11" VerticalAlignment="Center"
+                                                                                        <TextBlock Grid.Column="3" Text="{Binding Meta}" FontSize="11" VerticalAlignment="Center"
+                                                                                                   Margin="6,0,0,0"
                                                                                                    Foreground="{StaticResource KcapFaintBrush}"
                                                                                                    TextTrimming="CharacterEllipsis" />
-                                                                                    </StackPanel>
+                                                                                    </Grid>
                                                                                 </StackPanel>
                                                                             </Grid>
                                                                         </DockPanel>
